Create Stage 2 of 5
Description
In this stage, you'll learn how to convert each character of a string into a binary form. Any ASCII character has a size of 7 bits; in binary, the form is 0 and 1. For example, the character A has a decimal value of 65. The binary representation is 1000001. b is 98 in decimal or 1100010 in binary. The space character is 32 in decimal value or 0100000 in binary.

In this stage, you will learn to represent characters in binary form.